refactor(service): 纯移动拆分 admin_service.go（4409→642行）
